Allow serialization of Collections and Maps.
This does not play nicely with generics, however, because numbers will be
converted to Longs or Doubles, but the receiver may expect them to be of
a smaller type. Need to modify NUM to record which numeric type it is
storing.

Note: Three tests are still broken from the previous commit.