Support backreferences and superclass fields.
Backreferences allow the object graph to be correctly duplicated if the
same reference is present at multiple points, and fixes the
StackOverflowError problem that occurred when there was a cycle in the
graph.

Fields may now be fully qualified, and are when emitted. Before, the
parser could not find fields belonging to superclasses, and would throw
an exception when encountered. Now, unqualified fields are applied to the
first matching superclass, allowing simple hand-written usage to continue.

Note: Tests are now broken