Added an example project - A chat client/server.
Start the server, then start one or more clients. The clients each receive
the full chat history when they join. Two built-in commands can be sent
like messages: "\rename <new name>" and "\quit".

There is no built-in way to kill the server, so you need to Ctrl+C it (or
equivalent).