Moved some example code into the core library, and reorganized.
Created packages net.jonp.armi.comm.client and ...server for client-side
and server-side versions of the same code. Moved ClientSideCommunicator
and ServerSideCommunicator from net.jonp.armi.comm to the client and
server sub-packages, respectively.

Also, renamed and moved the SocketCommunicator classes from the example
project into the client and server package in the core library, as they
do not depend on any example code and are useful implementations of
communicators over Sockets.